Fixed code: renamed components
