+++ FrontEnd/src/components/AppFooter.vue
